[iOS] The completion handler in -handleKeyWebEvent:withCompletionHandler: is sometimes never called https://bugs.webkit.org/show_bug.cgi?id=214295 <rdar://problem/60539389>
Reviewed by Devin Rousso.

Source/WebKit:

This is a speculative fix for <rdar://problem/60539389>, wherein hardware key commands seemingly stop working in
a web page that is (presumably) otherwise responsive. It's possible that the bug exercises a scenario in which
the completion handler in `-[WKContentView handleKeyWebEvent:withCompletionHandler:]` is never invoked, which
subsequently leads to the keyboard task queue being backed up with key events.

This can happen in several ways. For instance, if the web process is swapped or terminates in the middle of
handling a key event, the key event queue will be cleared, but the UI process will still retain the (uncalled)
completion handler for that key event. Additionally, `WebPageProxy::handleKeyboardEvent` may not even have
attempted to propagate the event to the web process, in which case we shouldn't be saving the completion handler
and waiting for a response.

Test:   KeyboardInputTests.HandleKeyEventsInCrashedOrUninitializedWebProcess
        KeyboardInputTests.HandleKeyEventsWhileSwappingWebProcess

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::handleKeyboardEvent):

Make this return a `bool` indicating whether the key event was sent to the web process. If not, then we
should immediately invoke the completion handler in -handleKeyWebEvent:withCompletionHandler: below, instead of
stashing the Objective-C block and waiting for a response from the web process (which is presumably not
running).

* UIProcess/WebPageProxy.h:
* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView cleanUpInteraction]):
(-[WKContentView _cancelPendingKeyEventHandler]):

When the web process terminates or swaps in the middle of handling a key event, go ahead and invoke the key
event completion handler early with the queued event, since we aren't going to receive a response from the web
process anyways.

(-[WKContentView handleKeyWebEvent:withCompletionHandler:]):

Tools:

Add API tests to exercise the corner cases described in the WebKit ChangeLog.

* TestWebKitAPI/Tests/ios/KeyboardInputTestsIOS.mm:


Identifier: 227130@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264376 268f45cc-cd09-0410-ab3c-d52691b4dbfc