ASAN builds of WebKit should not freeze when an EXC_RESOURCE is delivered https://bugs.webkit.org/show_bug.cgi?id=214346
Patch by Tuomas Karkkainen <tuomas.webkit@apple.com> on 2020-07-16
Reviewed by Mark Lam.

When the WebContent process allocates 1000 MB of memory or starts 32
threads, it receives an EXC_RESOURCE. This requires iterating all the
mappings of the process. For the case of a process instrumented with
AddressSanitizer, the mappings total 20 terabytes, and cause the
process to "freeze" for several minutes.

Setting the sysctl debug.toggle_address_reuse to 1 causes the
EXC_RESOURCE mechanism to skip iteration of all the mappings.

With this patch, setting the environment variable
DISABLE_FREEZING_ON_EXC_RESOURCE will set the sysctl mentioned above
for WebContent and Networking processes.

* Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceMain.mm:

Identifier: 227201@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264450 268f45cc-cd09-0410-ab3c-d52691b4dbfc