CSS clip-path is applied to the <svg> root element in the view-box coordinates https://bugs.webkit.org/show_bug.cgi?id=214418
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2020-07-20
Reviewed by Darin Adler.

Source/WebCore:

We need to transfer the bounding box of the SVG element renderer from the
SVG coordinates system to the CSS coordinates before calculating the clip
path. Then we need to transfer the clip path back to the SVG coordinates
before applying it to the GraphicsContext. The GraphicsContext is already
transformed by the localToPrentTransform.

Test: svg/in-html/svg-view-box-css-clip-path.html

* rendering/svg/SVGRenderSupport.cpp:
(WebCore::SVGRenderSupport::clipContextToCSSClippingArea):

LayoutTests:

* svg/in-html/svg-view-box-css-clip-path-expected.html: Added.
* svg/in-html/svg-view-box-css-clip-path.html: Added.

Identifier: 227370@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264622 268f45cc-cd09-0410-ab3c-d52691b4dbfc