[JSC] eval?.() should be indirect eval https://bugs.webkit.org/show_bug.cgi?id=214568
Reviewed by Keith Miller.

JSTests:

* test262/expectations.yaml:
Mark two test cases as passing.

Source/JavaScriptCore:

eval?.() is specified as indirect eval, but (virtually) all implementations assumed it should be direct eval.
I raised this topic in today's TC39 meeting and we've decided to keep the spec as it is.

* parser/ASTBuilder.h:
(JSC::ASTBuilder::makeFunctionCallNode):
Don't use EvalFunctionCallNode for optional call of eval.


Identifier: 227379@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264633 268f45cc-cd09-0410-ab3c-d52691b4dbfc