[GStreamer][1.18] mediastreamsrc element hits assert in gst -core https://bugs.webkit.org/show_bug.cgi?id=214150
Reviewed by Philippe Normand.

Source/WebCore:

Removed track pointers from observers cause the object was
self-referencing a couple of times. Observers are 1-1 with the
source so there is no need to keep an owned reference and there is
no danger of incorrect dereferencing.

Turned finalize into dispose as the work is doing is more removing
the references with other objects, which what the dispose is
about. Called the parent on that, which as missing as well.

* platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp:
(WebKitMediaStreamTrackObserver::WebKitMediaStreamTrackObserver):
(WebKitMediaStreamObserver::WebKitMediaStreamObserver):
(webkitMediaStreamSrcConstructed):
(webkitMediaStreamSrcDispose):
(webkit_media_stream_src_class_init):

LayoutTests:

As result of this bug
fast/mediastream/MediaStream-removeTrack-while-playing.html begins
to crash in debug mode in an assertion. When a track is removed
the observer is not removed and when it dies there are null
objects in the observer weak hash set. Still, if you do a quick
patch to remove the observer, things can still crash as track
removal is racy and in release there' other crash. All this needs
to be taclked as a general approach to fix track removal.

* platform/glib/TestExpectations: Added
fast/mediastream/MediaStream-removeTrack-while-playing.html as
Pass Crash for release and crash for debug.


Identifier: 227394@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264648 268f45cc-cd09-0410-ab3c-d52691b4dbfc