[macOS] Datalist dropdown scrollbar position does not match visible region https://bugs.webkit.org/show_bug.cgi?id=214656
Reviewed by Wenson Hsieh.

Source/WebKit:

Currently, the content insets for the dropdown are set on an NSScrollView's contentView,
rather than the NSScrollView itself. On the latest macOS, this results in accurate state
within the NSScrollView, leading to an inaccurate scrollbar position. To fix, set the
content insets directly on the NSScrollView.

However, on older macOS, setting content insets on the NSScrollView causes the view to start
out slightly scrolled (by a value equal to the top inset). This issue is solved by preserving
the existing behavior (setting insets on the contentView) on older macOS.

Also, set the style of the embedded NSTableView to NSTableViewStyleFullWidth, to ensure the
view respects the insets we specify.

No new tests, as the change is purely visual and affects native UI.

* UIProcess/mac/WebDataListSuggestionsDropdownMac.mm:
(-[WKDataListSuggestionTableView initWithElementRect:]):
(-[WKDataListSuggestionsController initWithInformation:inView:]):

Source/WTF:

* wtf/PlatformHave.h: Define HAVE_NSTABLEVIEWSTYLE.


Identifier: 227515@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264783 268f45cc-cd09-0410-ab3c-d52691b4dbfc