Add a GNU parallel runner https://bugs.webkit.org/show_bug.cgi?id=214356
Patch by Angelos Oikonomopoulos <angelos@igalia.com> on 2020-07-24
Reviewed by Keith Miller.

At the moment, run-jsc-stress-tests uses make as a (local) job scheduling
engine (i.e. doesn't make use of the dependency tracking at all). However,
this causes problems for the distributed job scheduling we want to do when
using --remote.

If a remote is down, the tests will fail.
If a remote goes down during testing, the tests will fail.

There is no reason not to ignore remotes that are unavailable. What's more,
we should be able to reschedule jobs that were in the middle of execution when
a remote host goes down.

This patch tries to leverage GNU parallel as the execution engine to
hopefully transparently handle some of those failure scenarios.

* Scripts/run-javascriptcore-tests:
(runJSCStressTests):
* Scripts/run-jsc-stress-tests:

Identifier: 227539@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264813 268f45cc-cd09-0410-ab3c-d52691b4dbfc