[JSC] Do not use hardened Array for Intl supportedLocalesOf https://bugs.webkit.org/show_bug.cgi?id=214676
Reviewed by Mark Lam.

JSTests:

* stress/intl-suppored-locales-of.js: Added.
(shouldBe):

Source/JavaScriptCore:

We do not need to call getOwnPropertyNames & defineOwnProperty because hardening array of Intl.XXX.supportedLocalesOf is removed from the spec.
We should just return an array from bestFitSupportedLocales or lookupSupportedLocales, while this change is not observable to users (but it is better
for performance). This fully fixes https://github.com/tc39/ecma402/pull/278.

* runtime/IntlObject.cpp:
(JSC::supportedLocales):


Identifier: 227553@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264827 268f45cc-cd09-0410-ab3c-d52691b4dbfc