[iOS] Drag and drop does not preserve file names that contain periods https://bugs.webkit.org/show_bug.cgi?id=214701 <rdar://problem/66014009>
Reviewed by Tim Horton.

Source/WebCore:

Cocoa documentation does not make it clear whether the `-suggestedName` property on `NSItemProvider` should or
should not include the file extension. To avoid adding a redundant file extension, logic in
`WebItemProviderPasteboard` attempts to deduce whether the suggested name already contains an extension, and
only appends a file extension (taken from name of the original file) if the suggested name does not have one.

Unfortunately, since this logic only checks whether or not there is a period in the filename, it will
incorrectly avoid adding a file extension when the suggested name contains a period. To fix this, instead of
checking for the presence of a period, check to see whether the file extension of the suggested name is the
same, and append the extension if it isn't.

Test: DragAndDropTests.SuggestedNameContainsDot

* platform/ios/WebItemProviderPasteboard.mm:
(linkTemporaryItemProviderFilesToDropStagingDirectory):

Tools:

Add an API test to exercise both scenarios where the suggested name does and does not contain the extension.

* TestWebKitAPI/Tests/WebKitCocoa/WKAttachmentTests.mm:

Rebaseline an existing API test.

* TestWebKitAPI/Tests/WebKitCocoa/gif-and-file-input.html:

Adjust this test case to support handling multiple uploaded files.

* TestWebKitAPI/Tests/ios/DragAndDropTestsIOS.mm:
(TestWebKitAPI::TEST):


Identifier: 227674@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264962 268f45cc-cd09-0410-ab3c-d52691b4dbfc