[iOS] Unable to swipe on IMDB.com after long press on image https://bugs.webkit.org/show_bug.cgi?id=214968 <rdar://problem/66234421>
Reviewed by Wenson Hsieh.

Source/WebCore:

When a long press occurs on an <img>, a system drag interaction is initiated on iOS. In WebCore,
EventHandler::tryToBeginDragAtPoint() is called and a synthetic mouse event is produced, causing
handleMousePressEvent() to be called. Further down the call chain, dispatchPointerEventIfNeeded()
is called and a valid PointerEvent is generated in PointerCaptureController::pointerEventForMouseEvent()
with "pointerType" set to "mouse", even though there is already a touch interaction initiated.

We now check whether there are known touches before generating a PointerEvent for a MouseEvent.

In the case of IMDb, the page would keep track of "pointerdown" events to track whether a multi-touch
user gesture is in progress so that their slide shows can support two-finger zooming as well as
single-finger swiping. In the case of a long press, the second "pointerdown" event would trick
the code in thinking a zoom gesture was initiated and it never recovered.

Test: pointerevents/ios/long-press-yields-single-pointerdown-event.html

* page/PointerCaptureController.cpp:
(WebCore::PointerCaptureController::pointerEventForMouseEvent):

LayoutTests:

Add a test that triggers a long press gesture on an <img> and checks a single "pointerdown" event
was dispatched. Prior to the WebCore change in this patch, two events would be dispatched.

* pointerevents/ios/long-press-yields-single-pointerdown-event-expected.txt: Added.
* pointerevents/ios/long-press-yields-single-pointerdown-event.html: Added.


Identifier: 227788@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265088 268f45cc-cd09-0410-ab3c-d52691b4dbfc