The various AllowList options should be able to take the function name inline https://bugs.webkit.org/show_bug.cgi?id=215184
Reviewed by Saam Barati.

Right now when I use the various AllowList JSC options I almost
always only care about a single function. Right now you need to
create a file with that single name in it. That is inconvenient, so
this patch changes the behavior to treat the string as the
function name if no file at that path exists. I'm also
speculatively assuming fopen doesn't return ENOENT when it fails
due to sandboxing... I didn't feel like testing it because this is
a debug option.

* runtime/OptionsList.h:
* tools/FunctionAllowlist.cpp:
(JSC::FunctionAllowlist::FunctionAllowlist):

Identifier: 227979@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265299 268f45cc-cd09-0410-ab3c-d52691b4dbfc