Preload graphics drivers in Mac WebProcess https://bugs.webkit.org/show_bug.cgi?id=215183
Reviewed by Darin Adler.

Source/WebCore:

In newer versions of Mac OS, graphics drivers are no longer part of the shared cache due to
size restrictions. This can cause first render to be blocked by ~10 ms when we dlopen those
drivers. To work around this, we preload the drivers when prewarming the WebProcess.

* page/ProcessWarming.cpp:
(WebCore::ProcessWarming::prewarmGlobally):

Source/WTF:

Enable GPU driver preheating in versions of the OS that might not have the drivers in the dyld
shared cache due to size restrictions.

* wtf/PlatformEnableCocoa.h:


Identifier: 228096@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265418 268f45cc-cd09-0410-ab3c-d52691b4dbfc